Home page tags text added
BEACHES:
Manipal- Where the land greets the Ocean again. There is a beach here for all your moods. All the adieu is for you to decide on which beach you want to trod on and get a cab. Your Manipal Darshan is not done without the beaches.

MOUNTAINS:
The Arabic Sea on one side and the Western Ghats on the other. Manipal offers you a number of hills and mountain treks. You ought to camp on the heights of these mountains amidst the clouds and cold breeze of the Western Ghats.

TEMPLES:
This land is bestowed with tradition alongside the nature. The numerous temples, old and new, are an embodiment of the belief of the men of the land, as well as architectural excellence. Begin your Manipal Darshan with the Temples.

CLUBS:
The student town has an alluring night life with Clubs and Pubs all around the town. Dance your heart out with your friends at the amazing clubs in Manipal. Going out Club hopping late night is a must do in Manipal.

RESTAURANTS:
The land of Manipal is a buffet of cuisines hailing from different corners of the world. An Indian breakfast, Chinese lunch and Italian dinner: make your pick for your meals with Restaurants in and around Manipal.

ROAD TRIPS:
Got a days leave? Why sit in your rooms when you can have short road trips to nearby places. Discover a gallon of road trip ideas from Manipal and add them to your list for your next small vacation. Rent a bike and get going.